oldapi/requirements/requirements.txt
fastapi
uvicorn
python-dotenv
databases
